Apply `stderr` as new metric
Replace the `bootstrap_std` metric with `stderr`, a calulation of the standard error of the means using the Central Limit Theorem. We are able to compute the error directly using `stderr` (rather than using bootstrapping to estimate it) since the evaluation scores are means of numbers having finite variance.

Add `std` as a metric as well.