Implement provider for Google Vertex AI
This simply adds a new provider for Google Vertex AI. This uses a
different core python library to Google AI, which most people will use.
Vertex is used for more mature MLOps setups, which this should help us
support.

Although the core library is different (`vertexai` instead of
`google.generativeai`) the interface is very similar. Consequently, this
is largely a copy/paste from the `google.py` provider, with relevant
parts fixed up.

Note that I refrained from DRYing up `safety_settings` in a shared
utility because the types are different, but also Vertex supports
`HARM_CATEGORY_UNSPECIFIED` whereas Google AI doesn't. Although minor,
it seems reasonable to assume these may diverge in other ways in the
future, but I'm happy to extract this to a shared util if required!