Log View Auto False Improvements
Improve overall behavior when the user has the log view auto setting disabled. When this setting is disabled, we now:

- Continue to not open the viewer for logs when they complete

- But if the viewer is open, we will attempt to refresh it, but only if the view isn’t focused (the idea being that if the user is actively working in the viewer, we will leave the state alone). If the viewer is focused, we’ll just refresh the list of logs so the new log appears in the list

Note that this behavior will only occur if the viewer is opened to the same log_dir as the task which just completed. We will not modify the log viewer if log directories are changed.