Add proposed `exact` and `f1` scorers.
`exact()`
Scorer which will normalize the text of the answer and target(s) and perform an exact matching comparison of the text. This scorer will return `CORRECT` when the answer is an exact match to one or more targets.

`f1()`
Scorer which computes the `F1` score for the answer (which balances recall precision by taking the harmonic mean between recall and precision).