Remove restrictive `open with` regex
We already ensure that the file is an eval log before we open the viewer, so allowing the command to appear will make it more useful to users who rename log files.

The command is in the middle of the context menu, so it doesn’t add a lot of distraction if we occassionally show it for non-inspect log files.