Allow more complex metrics (#378)
* Allow more complex metrics

Allow more complex metrics like:

```
@scorer(
  metrics=[
  {
    "score1": [mean(), stderr()],
    "score2": [mean(), stderr()]
  },
  f1_metric()]
)
```

The above show a list of metrics that will be applied to a score whose value is a dictionary. The dictionary defined in the metric will be applied to the dictionary score (by key). The `f1_metric` in the example will be passed the complete score.value dictionary and need to compute a score based upon that dictionary value.

* Fix issue with mypy

* Add to changelog

* fix lint

* Fix mypy complaint about test

---------

Co-authored-by: jjallaire <jj.allaire@gmail.com>