Prune Sample Reductions
When returning `header_only` we’re trying to provide macro level performance and metadata while omitting sample specific data which inflates the size the of the data. We are currently leaving `sample_reductions` when providing `header_only`, which can inflate the size of the headers file since it includes an entry for each sample (bad, for example, when model grading and each reduced sample carries an explanation from the model).