Improve answer detection in multichoice scorer (#730)
fixes #721

In cot responses particularly, the string `answer:` could appear in the reasoning tokens, which in some cases confuses the extraction regex for determining the answer.

This fix takes a first pass using a more strict search, looking for a line which contains `answer: A`, using that as the answer if present. If no such line is present, we will use the previous search to determine the answer.