Improve display of large number of config/parameters in log viewer (#816)
* Add support for styles, recompute collapse is children change

* Add expandable panel to scores

* updated js

* add changelog

* Application lineclamp to sidebar hyperparams

prettier

* Better compute selected scorer

We depend upon ‘selectedScorer’ when rendering score values, but that previously was coming from results. This meant that if results weren’t present because the evaluation was still running, the score value wouldn’t be displayed.

This updates our default scorer logic to try to use the results to compute the scorers and default scorer, but if that it isn’t present, it will compute it using the first sample instead.