Properly handle ctrl+c
We previously used `gather` to await the completion of the tasks. gather auomatically cancels pending tasks when it is interupted, but our new implementation wasn’t handling errors (thinking this was correct since teh first throw exception would be returned).

We do need to raise the first thrown exception, but in this case, we also need to cancel pending tasks since we are exiting. This change accumulates the tasks themselves, then cancels any tasks that aren’t done when an error is raised (then re-raises the error).