Correct render of scores with partial metrics
Scorers that don’t map every dictionary value to a metric, like:

```
@scorer(metrics={"basic": [mean(), stderr()], "cool": [accuracy(), stderr()]})
def subset_scorer():
    async def score(state: TaskState, target: Target):

        return Score(value={
            "basic": 1.0 * random.random(),
            "fancy": 0.9 * random.random(),
            "cool": 1.1 * random.random(),
            "uncool": 2.1 * random.random()
        })
    return score
```

Will result in a score value being a dictionary, but a scorer name for metric. We previously wouldn’t render individual scores unless all the score values are used in metrics, this changes it to apply this logic if any of the keys are used in metrics.