Support dynamic metric names (#1003)
* Support dynamic metric names

The first implementation assumed that metric names would be static through the eval, but there are cases (e.g. agentharm) where keys may be changed (for example, a category level metric). This removes the assumption that metric names are static.

We use metric names to pick widgets to update (so we update only the value, not recomputing the whole grid constantly) - when we encounter a metric that we don’t have a value widget for, this means the grid is stale and needs to be re-computed.

* Update CHANGELOG.md

---------

Co-authored-by: jjallaire <jj.allaire@gmail.com>