remove use of aiofiles
We currently use aiofiles for sandbox read_file implementations as well as for writing auto-compose files. These are all local file reads and they are for small files so in most cases this will make read performance considerably worse (as it schedules the read/write onto a thread pool).

Further, we are attempting to eliminate any possibility of threading deadlocks, and this represents another point of exposure to them.