Truncate Tracebacks
If tracebacks exceed 1MB, we will truncate them. In this case, we will preserve at least 2048 characters of the traceback itself, and will truncate the error_message itself (which is the issue in the case of attempting to access a dict using a very large object as a key).

When we truncate the traceback we downgrade ansi to plain text.

tweak algorithm