Avoid nesting virtualized transcripts.
Virtualized transcripts interact with the scroll parent, so when they are nested, they compete with the root transcript for scroll interactions. This makes inner transcripts non-virtualized. For the time being this is fine, eventually I’ll make them virtualized but without any of the scroll interaction and/or nested height computation issues.