Fix tests
This changes tests that are comparing logs (e.g. generating a log, writing it, then re-reading it) to compare the serialized representations rather than the in-memory representations. This will ensure that the persisted form must be identical without requiring that the in memory form be identical.

This is specifically to support memory only properties that might exist, for example, the `id_` field on the events, which exist primarily to deduplicate events that we send to the client (and aren’t persisted and are intended to be transient, consequently each in memory representation of the log gets its own ids since they are never serialized).