Remove `chdir` option from `@task` (#1660)
This had previously been disabled by default. This change means that tasks can no longer opt-in to changing their working directory during execution (this was an undocument option that we left in in the case that people wanted this behavior back, but we haven't subsequently received any requests for it.