Treat threshold as inclusive minimum
I reviewed inspect_evals usage of the pass_at reducer and each example will continue to report correct scores (they’re all using Correct/Incorrect with threshold of 1). This would be a breaking change for users who pass a threshhold expecting it to be an exact match, but I think that is a bug since threshhold implies a minimum rather than an exact value.