Prefer range request to determine eval file size
HEAD requests may be fulfilled using compression and this results in a compressed file size rather than an uncompressed file size which is needed for downstream range requests. (accept-encoding is also a way to control this, but the browser controls this, so we can’t set it for the HEAD request).

The downside of this approach is an extra range request (for 1 byte) which will always do if ranges are allowed (before falling back to HEAD requests).