Thin summary inputs
Its possible for text inputs to be quite large (for example, you could imagine the input being the result of another model trajectory, which itself could be megabytes to hundreds of megabytes). The summary is intended to be a small / compact representation of the sample, so carrying a massive input isn’t desirable.

Not truncating results in the JSON encoded summaries within the eval file being hundreds of megabytes which overwhelms reasonable JSON deserialization on the client.