code_viewer: handle non-string arguments
Sometimes models call functions with incorrect argument types. (This is
particularly notable in the newly-released gpt-oss, which has a strong
tendency to pass a list of strings rather than a single string to the
bash command.)

The actual function call logic handles this gracefully by returning a
string describing the argument parsing error, so the model can correct
it.

But the viewer was crashing when the argument type was incorrect.
Instead of crashing, it now converts whatever it gets to a string using
`str`.