Fix issues resolving scorers
1) Add return type annotations to the tests.
When loading scorers, we use the return type annotation to verify that the function is a scorer (this is because in the very early days of Inspect, we supported decorating functions as scorers directly rather than factories as scorers).

2) Use resolve_scorers
This function will resolve the scorers directly from the logs. Updated signature to permit passing on the log, in which case it will properly resolve the scorers from the log, if possible.