add score editing (#2433)
* add score editing

* remove unnecessary code in score_value

* revert changes to metrics_from_log_header()

* fix call to eval_results()

* rename _edit.py to _score.py

* move recompute_metrics() and create alias

* update CHANGELOG.md after rebase

* Fix issues resolving scorers

1) Add return type annotations to the tests.
When loading scorers, we use the return type annotation to verify that the function is a scorer (this is because in the very early days of Inspect, we supported decorating functions as scorers directly rather than factories as scorers).

2) Use resolve_scorers
This function will resolve the scorers directly from the logs. Updated signature to permit passing on the log, in which case it will properly resolve the scorers from the log, if possible.

* fix tests

* remove accidentally committed file

---------

Co-authored-by: dragonstyle <cteague@gmail.com>