Rework treeify of spans
The previous implementation only treeified correctly for ordered Spans (e.g. begin spans followed by events followed by an end span), which is common, but is incorrect. In async cases, events can be interleaved.

To solve this, spans and events carry span_id’s which specifies what span they are a part of. This refactors treeification in the viewer to properly treeify using the span_ids.