Inspect View: Fix Issue with Incorrect Log Display (#2543)
* Move unloading log to more logical location

This shouldn’t really change the behavior but is just more understandable here.

* Don’t select a log if one wasn’t selected

When a new log is received, we end up calling ‘refresh’ which adds a log to the list. In the case that no log was currently selected, we would treat the first log in the old list as if it were selected, then restore that selection when the refresh was complete.

This means that when subsequent pages log, they will have a loaded version of the previous log loaded, resulting in potential display issues.

---------

Co-authored-by: jjallaire <jj.allaire@gmail.com>