Anthropic: Increase default `max_tokens` to 32,000 for Claude 4 models (#2675)
Anthropic requires you to explicitly specify max_tokens (most others set it to the maximum allowable output tokens for the model). Claude 4 models currently have maxes of 32k (Opus 4 and 4.1) or 64k (Sonnet 4 and 4.5 and Haiku 4.5). Claude 3 models max at 4k or 8k, w/ 3.7 Sonnet being capable of128k w/ a special header).

Therefore, we use 4k as the default for Claude 3 and 32,000 as the default for everything else.