Correct `eval-set` route `log_dir` handling (#2725)
* Correct `eval-set` route `log_dir` handling

The `eval-set` route is passed a directory as a query param (this is the subdirectory to check for eval-sets). It should also be passed a `log_dir` (optionally), which can server as the base log directory for the query. This adds both parameters to the request on the client side and adds proper server side handling of the request (properly joining the `log_dir` with the `dir` and so on).

* don’t prepend slashes for empty base directory

* Correct grid layout for LogsPanel