Improve Client Replication Performance (#2837)
* Remove extraneous log message

* Set keep-alive to 15

For default server there was no keep-alive, for fastapi we were using default of 5. (This didn’t really move the needle on perf but it still righteous).

* Display sample count in footer

* Add bulk writing of previews

* Improve concurrency in replication work

Previous replication was working in batches rather than just saturating a pool of concurrent workers. This converts the workqueue to a concurrent queue which will improve throughput of the queue.